+++ formula/src/main/java/com/instacart/formula/Formula.kt
@@ -35,8 +35,8 @@ interface Formula<in Input, State, out Output> : IFormula<Input, Output> {
     /**
      * The primary purpose of evaluate is to create an [output][Evaluation.output]. Within
      * this method, we can also [compose][FormulaContext.child] child formulas, handle
-     * callbacks [with data][FormulaContext.eventCallback] or [without data][FormulaContext.callback],
-     * and [respond][FormulaContext.updates] to arbitrary asynchronous events.
+     * callbacks [FormulaContext.onEvent], and [respond][FormulaContext.updates] to
+     * arbitrary asynchronous events.
      *
      * Evaluate will be called whenever [input][Input], [internal state][State] or child output changes.
      *