+++ formula/src/main/java/com/instacart/formula/internal/Callbacks.kt
@@ -1,8 +1,7 @@
 package com.instacart.formula.internal
 
 internal class Callbacks {
-    private var callbacks: SingleRequestMap<Any, Callback>? = null
-    private var eventCallbacks: SingleRequestMap<Any, EventCallback<*>>? = null
+    private var callbacks: SingleRequestMap<Any, Callback<*>>? = null
 
     private fun duplicateKeyErrorMessage(key: Any): String {
         if (key is String) {
@@ -13,44 +12,37 @@ internal class Callbacks {
         return "Callback $key is already defined. Are you calling it in a loop or reusing a method? You can wrap the call with FormulaContext.key"
     }
 
-    fun initOrFindCallback(key: Any): Callback {
+    fun initOrFindCallbackT(key: Any): UnitCallback {
         val callbacks = callbacks ?: run {
-            val initialized: SingleRequestMap<Any, Callback> = mutableMapOf()
+            val initialized: SingleRequestMap<Any, Callback<*>> = mutableMapOf()
             this.callbacks = initialized
             initialized
         }
 
         return callbacks
-            .findOrInit(key) { Callback(key) }
+            .findOrInit(key) { UnitCallback(key) }
             .requestAccess {
                 duplicateKeyErrorMessage(key)
-            }
+            } as UnitCallback
     }
 
-    @Suppress("UNCHECKED_CAST")
-    fun <UIEvent> initOrFindEventCallback(key: Any): EventCallback<UIEvent> {
-        val eventCallbacks = eventCallbacks ?: run {
-            val initialized: SingleRequestMap<Any, EventCallback<*>> = mutableMapOf()
-            this.eventCallbacks = initialized
+    fun <UIEvent> initOrFindCallbackT(key: Any): Callback<UIEvent> {
+        val callbacks = callbacks ?: run {
+            val initialized: SingleRequestMap<Any, Callback<*>> = mutableMapOf()
+            this.callbacks = initialized
             initialized
         }
 
-        return eventCallbacks
-            .findOrInit(key) { EventCallback<UIEvent>(key) }
+        return callbacks
+            .findOrInit(key) { Callback<UIEvent>(key) }
             .requestAccess {
                 duplicateKeyErrorMessage(key)
-            } as EventCallback<UIEvent>
+            } as Callback<UIEvent>
     }
 
     fun evaluationFinished() {
         callbacks?.clearUnrequested {
-            it.callback = {
-                // TODO log that disabled callback was invoked.
-            }
-        }
-
-        eventCallbacks?.clearUnrequested {
-            it.callback = {
+            it.delegate = {
                 // TODO log that disabled callback was invoked.
             }
         }
@@ -58,17 +50,10 @@ internal class Callbacks {
 
     fun disableAll() {
         callbacks?.forEachValue {
-            it.callback = {
+            it.delegate = {
                 // TODO log that event is invalid because child was removed
             }
         }
         callbacks?.clear()
-
-        eventCallbacks?.forEachValue { entry ->
-            entry.callback = {
-                // TODO log that event is invalid because child was removed
-            }
-        }
-        eventCallbacks?.clear()
     }
 }