+++ formula/src/test/java/com/instacart/formula/ChildMessageWithParentStateChange.kt
@@ -3,7 +3,7 @@ package com.instacart.formula
 object ChildMessageWithParentStateChange {
     fun formula() = HasChildFormula(MessageFormula()) { state ->
         MessageFormula.Input(
-            messageHandler = eventCallback {
+            messageHandler = onEvent<Int> {
                 transition(state + 1)
             }
         )