+++ formula/src/test/java/com/instacart/formula/MultipleChildEvents.kt
@@ -2,7 +2,7 @@ package com.instacart.formula
 
 object MultipleChildEvents {
     fun formula() = HasChildFormula(MessageFormula()) {
-        MessageFormula.Input(messageHandler = eventCallback {
+        MessageFormula.Input(messageHandler = onEvent<Int> {
             transition(it)
         })
     }